==> 05-results.Rmd <==
# Results
The results chapter is broadly divided into two sections - one for each of the goals of this project. The first section will document all preliminary findings related to the sleep measurements and the procedure of defining a metric for quantifying quality of sleep.

## Metric Formulation
### Preliminary Observations
Formulating a metric requires an extensive analysis of the data to identify correlated variables. This is to avoid multicollinearity while defining the metric. Below is a correlation heatmap we generated for variables that do not have the format of a time stamp (for example hh:mm). We made the following observations - 

(1) Latency, defined as the time taken for the user to fall asleep after he gets into bed, does not seem to be correlated with anything 
(2) Total sleep time (TST) is not correlated with neither Number of Awakenings and Average Awakening Length. One would expect TST to be negatively correlated to both but that is not true 
(3) Movement, Fragmentation and Sleep Fragmentation index are highly correlated. It is not surprising as sleep fragmentation index is defined as the ratio of the two
(4) Wake After Sleep Onset (WASO) is strongly positively correlated to almost every variable except for Efficiency
(5) Efficiency, defined as percentage of sleep time on total sleep in bed, surprisingly is not correlated to Total Minutes in Bed

Note: all coefficients are statistically significant with a default critical value of 0.05

```{r}
library(Hmisc)
library(corrplot)
library(gplots)
library(purrr)
load("./data/RData/data_MMASH_cleaned.RData")
sleep_data <- read.csv('data/raw/multilevel-monitoring-of-activity-and-sleep-in-healthy-people-1.0.0/main_sleep_transformed.csv')
# adding user id
sleep$user.id <- as.numeric(unlist(map(strsplit(sleep$user, split = "_"), 2)))
sleep <- sleep[order(sleep$user.id),]

sleep_data <- as.data.frame(t(sleep_data))
colnames(sleep_data) <- sleep_data[1, ]
sleep_data <- sleep_data[-1, ]

#removing variables with time entries of format hh::mm
sleep_subset <- subset(sleep_data, select = -c(1,3,5))

# converting all values that are not time readings to numeric type
sleep_subset[,4:length(sleep_subset)] <- sapply( sleep_subset[,4:length(sleep_subset)], as.numeric)
mydata <- sleep_subset[,4:length(sleep_subset)]

mydata.rcorr = rcorr(as.matrix(mydata))
mydata.coeff = mydata.rcorr$r
mydata.p = mydata.rcorr$P
palette = colorRampPalette(c("darkblue", "white", "darkred")) (20)
heatmap.2(x = mydata.coeff, symm = TRUE, labRow = abbreviate(rownames(mydata.coeff), minlength = 24), labCol = abbreviate(colnames(mydata.coeff), minlength = 24), col = palette, margins=c(10,16), srtCol=45,main = "heatmap", density.info="none", trace="none")

mydata['user'] <- c(1,2,3,4,5,6,7,8,9,10,12,13,14,15,16,17,18,19,20,21,22)
```

### Distribution of Variables
There are 3 clusters in the heatmap (if we cut across the tree right below the highest line connecting efficiency and another cluster) which are consistent with the observations we made. To avoid multicollinearity, we picked variables from each of the clusters depending on the strength and significance of the correlation coefficient. We denote our metric by M. Intuitively, it should be higher for users who are able to sleep through the night and do not have any trouble falling asleep. Keeping this in mind, we came up with the following associations - 

(1) M $\alpha$ Total Sleep Time (TST)
(2) M $\alpha$ $\frac{1}{SleepFragmentation Index}$
(3) M $\alpha$ $\frac{1}{Latency}$
(4) M $\alpha$ $\frac{1}{AverageTimeSpentAwake}$

We combined the above associations as follows - 

\begin{align*}
M \alpha (\frac{TST}{AverageTimeSpentAwake} + \frac{1}{SleepFragmentationIndex}).\frac{1}{Latency}
\end{align*}

The units of the right hand side are "per hh:mm" (because of addition of latency). Average Time Spent Awake is the product of Number of Awakenings and Average Awakening Length. Let the ratio of TST and Average Time Spent Awake be called sleep ratio.

Next, we quantified the distribution of variables involved in defining M. We wanted to include informative variables instead of the ones that remain constant across the cohort. Due to the small sample size, any analysis of such sort (constructing a metric) will have a selection bias and so the conclusions made throughout this section may not hold for a bigger cohort.

The box plot below shows the distribution of the Sleep Ratio and Sleep Fragmentation Index. The values for both ratios seems well distributed with no outliers.

```{r}
library(dplyr)
mydata['Average Time Awake'] = mydata$`Number of Awakenings`*mydata$`Average Awakening Length`
mydata['Sleep Ratio'] = mydata$`Total Sleep Time (TST)`/mydata$`Average Time Awake`
df <- dplyr::select(mydata,c('Sleep Ratio','Sleep Fragmentation Index'))

ggplot(stack(df), aes(x = ind, y = values)) + geom_boxplot() +labs(x = "", y = "Ratio Values", title = "Distribution of Sleep Ratio and Sleep Fragmentation Index in Cohort") 
```

Since Latency is not a ratio, we visualized it separately as a Cleveland dot plot. There are a lot users with latency time of 0 mins which is a problem since M is inversely proportional to latency. So, we added a dummy term to latency and replaced proportionality with equality (we assume proportionality constants to be 1 - discussed in Conclusions chapter)


```{r}
ggplot(df, aes(mydata$Latency, reorder(mydata$user, mydata$Latency))) + 
  geom_point() + 
  # scale_x_continuous(breaks = scales::pretty_breaks(n = 17)) +
  labs(y = "User ID", x = "Time in Mintues", title = "Latency per User") +
  # geom_segment(aes(yend = reorder(mydata$user, mydata$Latency)), xend = 0, colour = "grey50") +
  # theme(panel.grid.major.y = element_blank())
 theme_linedraw()
```

We define the sleep quality of a user as - 

\begin{align*}
M = (\frac{TST}{AverageTimeSpentAwake} + \frac{1}{SleepFragmentationIndex}).(\frac{1}{Latency + 1})
\end{align*}

In the case that other variables are 0, we simply replace them by 1. We did not add dummy terms as the none of the users in our cohort have $TST = 0$ or $Number of Awakenings = 0$. M is always positive as all the variables involved are non-negative.

### Distribution of Sleep Quality Metric

We calculated the sleep quality metric for all users and visualized the results as a boxplot. The lowest value is 0.5902 corresponding to user 2 and the largest is 16.33 corresponding to user 12, who is also an outlier. We are interested in investigating users that fall below the lower hinge (users 2, 3, 8, 10, 16) and above the upper hinge (users 1, 12, 14, 20, 21).

```{r}
library(tidyverse)
# calculating the metric
mydata['M'] = ((mydata$`Total Sleep Time (TST)`/mydata$`Average Time Awake`)+(1/mydata$`Sleep Fragmentation Index`))*(1/(mydata$Latency + 1))

#adding M to sleep data
sleep$M <- c(rep(mydata$M[1],2), mydata$M[2:length(mydata$M)])

# ggplot(mydata, aes(`user`, `M`)) + geom_point() + scale_x_continuous(breaks = scales::pretty_breaks(n = 17)) + labs(x = "User ID", y = "Units (per minute)", title = "Sleep Quality of each User") + theme_bw()

ggplot(stack(dplyr::select(mydata, c('M'))), aes(x = ind, y = values))  + geom_boxplot() + geom_jitter() +labs(x = "Sleep Quality", y = "Units (per minute)", title = "Distribution of Sleep Quality") 

fivenumnames <- c("min", "lower-hinge", "median", "upper-hinge", "max")
l <- mydata$M
fivenum(l) %>% set_names(fivenumnames)
```

Having defined M, we wanted to see if it is capturing the quality of sleep. We used `Efficiency` to evaluate its accuracy. Given the definition of Efficiency, it should be positively correlated to M.

To test the relationship, we generated a log-log scale plot. We used a log10 scale to normalize the values and to see how the magnitudes of the two variables are related. The regression model is statistically significant with a p-value of 3.681e-05. Positive slope suggests that as M increases, efficiency also increases. However, the adjusted R square value is 0.6 which means that the model captures less than 70% of the variance. 

In summary, the metric defined by the formula above seems to be capturing the information that directly affects the sleep quality of a user. Moreover, the statistical significance combined with the positive slope of the regression model and line prove to be a sanity check for our metric.

```{r}
library(ggpubr)
ggplot(mydata, aes(x = log(`M`, 10), y = log(`Efficiency`, 10))) + geom_point() + geom_smooth(method="lm") + labs(title = "Efficiency vs Sleep Quality (log10 scale)") + stat_regline_equation(label.x=-0.3, label.y=1.99) + stat_cor(aes(label=..rr.label..), label.x=-0.3, label.y=1.98)

# model <- lm(log(mydata$Efficiency, 10)~log(mydata$M, 10), header = T)
# summary(model)
```


## Establishing Direct & Indirect Effects on Sleep Quality
In this section, we tried to identify variables that are correlated to sleep quality in some way. The relationships identified cannot be interpreted to be causal as correlation doesn't imply causation.

This section is divided further into subsections for each file in the dataset. Due to redundancy in data, we analyzed only 4 out of the 7 files - sleep records, user information, activities performed and saliva samples.

### Does Bed Time Affect Sleep Quality?
The practice going to bed early and rising early is encouraged for better sleep and health. We wanted to test this hypothesis using our metric. 

We generated a scatterplot and concluded that bed time does not have affect sleep quality. There is no linear relationship; some participants who went to bed later than "usual" have a higher sleep quality than most participants who went to bed early. Similar conclusion can be made for the time a participant got out of bed. The data has high variance and so we cannot conclude that bed times and sleep quality are related.

This can be justified as sleep quality is most likely to be affected by how tired the user feels and total time they slept the day before. 

```{r}
# formatting sleep_data
sleep_data$`Out Bed Date`[sleep_data$`Out Bed Date` == 1] <- 2
sleep_data$In.Bed.Time = string2time(sleep_data$`In Bed Time`, format = "%H:%M") + lubridate::days(sleep_data$`In Bed Date`)
sleep_data$Out.Bed.Time = string2time(sleep_data$`Out Bed Time`, format = "%H:%M") + lubridate::days(sleep_data$`Out Bed Date`)
sleep_data$M = mydata$M

temp = sleep_data %>% select(`In.Bed.Time`, `Out.Bed.Time`, M) %>% gather("Bed.Time", "Time")
rownames(sleep_data)
temp = sleep_data %>% 
  select(`In.Bed.Time`, `Out.Bed.Time`, M) %>% 
  mutate(user = sapply(rownames(sleep_data),simplify =TRUE, function(x)as.numeric(str_remove(x, "user_")))) %>% 
  gather("Bed.Time", "Time", 1:2) 

temp %>% 
  ggplot(aes(x = Time, y = M))+
  geom_point(aes(color = Bed.Time), alpha= 0.5) +
  labs(title = "Sleep Quality vs Bed Time", x = "Time (hh:mm)", y = "Sleep Quality (M)", col = "Color") + 
  scale_color_manual(labels = c("In Bed Time", "Out Bed Time"), values = c("red", "blue")) +
  theme_bw()
  # geom_line(aes(group = user), color = "grey", linetype = 2)

```


### Does Age and Build Affect Sleep Quality?

It has been seen that as a person ages their quality of sleep deteriorates. We analyzed the relationship between age and our sleep quality metric to see if such a pattern emerges. A general trend appeared in the scatterplot, excluding certain outliers. M seems to be linearly related to age till a threshold of 30 years. We saw a sharp decline in sleep quality for users older than 30 years. Moreover, sleep quality remained more or less constant beyond the threshold and is much lower compared to young adults. It is important to keep in mind that sample ages are extremely skewed and so similar pattern might not emerge with the addition of more middle aged users. 

Next, we wanted to see if taller people sleep better or not. Similar to the ages, height of users is also skewed. However, we still see close to a linear relationship with a negative slope between M and height till 190 cm. One reason could be, it is easier for short people to get comfortable in bed and that helps them sleep better. We did not categorize users as outliers due to small sample size and low diversity.

We then moved on to see if weight affects sleep quality. As expected, there seems to be no direct relationship between the two. The scatterplot shows the spread of the data which doesn't have any discernible pattern. Users with weight 70 kg and 80 kg have a range of values for sleep quality.

Since the authors of the data only studied males, we cannot conclude if one gender is inclined to have better sleep quality than another. We refrained from adding more geneder levels in the plot to be politically correct.

```{r, figures-side, fig.show="hold", out.width="50%"}
user_info$user.id <- as.numeric(unlist(map(strsplit(user_info$user, split = "_"), 2)))
user_info <- user_info[order(user_info$user.id),]
user_info_temp <- user_info[user_info$user.id != 11,]
user_info_temp$M <- mydata$M
user_info_temp <- user_info_temp[order(user_info_temp$Age),]
user_info_temp[user_info_temp == 0] <- NA

par(mfrow=c(1,3))  
ggplot(user_info_temp, aes(`Age`,`M`)) + geom_point() + labs(title = "Sleep Quality vs Age of Participants",y = "Sleep Qualtiy") + theme_bw()
ggplot(user_info_temp, aes(`Height`,`M`)) + geom_point() + labs(title = "Sleep Quality vs Height of Participants") +theme_bw()
ggplot(user_info_temp, aes(`Weight`,`M`)) + geom_point() + labs(title = "Sleep Quality vs Weight of Participants",y = "Sleep Qualtiy") + theme_bw()
ggplot(user_info_temp, aes(`Gender`,`M`)) + geom_point() + labs(title = "Sleep Quality vs Gender of Participants") +theme_bw()
```


### Which Activities are Deterimental to Sleep Quality?

From the `Activity.csv`, we explored the correlation between the sleep quality metric and the total time each user spent on each activity throughout the day. We generated a heatmap of the time each user spends on each activity. The blank spaces mean that the user did not engage in that activity at all. Initially, we were curious about alcohol and caffeine consumption but the amount of data was too less to analyze. We decided to discuss screen usage and heavy activities. 

```{r}
activity_duration %>% ggplot(aes(x = factor(user, levels = paste0("user_",c(1:22)), labels = c(1:22)), y = factor(category), fill = total)) +
  geom_tile(color = 'black', lwd = 0.4, linetype = 1) +
  scale_x_discrete(labels = abbreviate) +
  theme(panel.background = element_blank() ) + 
  scale_fill_distiller(palette = "Blues", direction = 1) +
  labs(title = "Time spent on each activity", x= "User ID", y = "Activity", fill = "Seconds Spent\non Activity")
```

The correlation between the total time each user spent on each activity and the sleep quality metric is shown below. It can be seen that M is positively correlated to screen usage which seems counterintuitive. So, we extracted the p-values corresponding to the correlation matrix and found that none of the coefficients (including M) were statistically significant.

```{r}
df_act = activity_duration [-2] %>% pivot_wider(names_from = category, values_from = total, values_fill = 0)
df_act['user'] = lapply(df_act['user'],function(x) as.numeric(str_remove(x, "user_")))

# filter out user 11
df_act = df_act %>% arrange(user) %>% filter(user!=11)
df_act['M'] = mydata$M
df_act.rcorr = rcorr(as.matrix(df_act[-1]))
df_act.coeff = df_act.rcorr$r
df_act.p = df_act.rcorr$P

heatmap.2(x = df_act.coeff, symm = TRUE, key = TRUE,labRow = abbreviate(rownames(df_act.coeff), minlength = 24), labCol = abbreviate(colnames(df_act.coeff), minlength = 24), col = palette, margins=c(8,10), srtCol=45, main = "heatmap", density.info="none", trace="none")
```

To dig deeper, we looked at the relationship between the amount off-screen time the user has right before sleeping and the sleep metric. We expected to have a positive linear relationship. However, from the regression lines below we see that the relationship is neither linear nor positive. Plot with averaged small and large off-screen time is comparatively closer to our expectations but the standard errors are too large to ignore.

```{r, fig.show="hold", out.width="50%"}
# get the latest time using screen before sleep outset time
temp = sleep %>% select(Onset.Time, user) %>% group_by(user) %>% top_n(1, desc(Onset.Time)) 
scrn_usg_time = activity %>% 
  filter(Activity == 8|Activity == 9) %>% 
  left_join(temp, by = 'user') %>% 
  filter(End < Onset.Time) %>% 
  group_by(user, Activity) %>% 
  top_n(1, End) %>% 
  select(Activity, End, user, Onset.Time) %>% 
  pivot_wider( names_from = Activity, values_from = End)

colnames(scrn_usg_time) = c('user', "Onset.Time", "small scr.usg", "large scr.usg")
scrn_usg_time['user'] = lapply(scrn_usg_time['user'],function(x) as.numeric(str_remove(x, "user_")))
scrn_usg_time = scrn_usg_time %>% arrange(user) %>% filter(user!=11) 
# user 13 and 22 is missing screen usage information
# user 18 didn't use screen before sleep
scrn_usg_time['M'] = mydata[mydata$user != 13 & mydata$user != 22 & mydata$user != 18,]$M
scrn_usg_time[is.na(scrn_usg_time)] = as.POSIXct('2021-12-13 00:00:00')

scrn_usg_time['scr.usg'] = pmax( scrn_usg_time$`small scr.usg`, scrn_usg_time$`large scr.usg`)

scrn_usg_time['large.scr.before.sleep'] = lapply(scrn_usg_time['Onset.Time'] - scrn_usg_time['large scr.usg'], function(x) as.double.difftime(x)*60)

scrn_usg_time['small.scr.before.sleep'] = lapply(scrn_usg_time['Onset.Time'] - scrn_usg_time['small scr.usg'], function(x) as.double.difftime(x))

scrn_usg_time['scr.before.sleep'] = lapply(scrn_usg_time['Onset.Time'] - scrn_usg_time['scr.usg'], function(x) as.double.difftime(x))

# ggplot(scrn_usg_time, aes(`large scr.usg`, M)) + geom_point() + labs(title = "Sleep Quality vs Screen Usage") +theme_bw()
# ggplot(scrn_usg_time, aes(`small scr.usg`, M)) + geom_point() + labs(title = "Sleep Quality vs Screen Usage") +theme_bw()
# ggplot(scrn_usg_time, aes(`scr.usg`, M)) + geom_point() + labs(title = "Sleep Quality vs Screen Usage") +theme_bw()
ggplot(scrn_usg_time, aes(scr.before.sleep, M)) + 
  geom_point() + 
  labs(title = "Sleep Quality vs Screen Usage") +
  theme_bw() +
  xlab("Off-screen Mimute Before Sleep") + 
  ylab("Sleep Quality(M)") + 
  # scale_x_continuous(trans='log2')+
  geom_smooth(method=lm) 

ggplot(scrn_usg_time, aes(large.scr.before.sleep, M)) + 
  geom_point() + 
  labs(title = "Sleep Quality vs Large Screen Usage") +
  theme_bw() + 
  xlab("Off-screen Mimute Before Sleep") + 
  ylab("Sleep Quality(M)") +  
  # scale_x_continuous(trans='log2') + 
  geom_smooth(method=lm) 

ggplot(scrn_usg_time, aes(small.scr.before.sleep, M)) + 
  geom_point() + 
  labs(title = "Sleep Quality vs Small Screen Usage") +
  xlab("Off-screen Mimute Before Sleep") + 
  ylab("Sleep Quality(M)") + 
  theme_bw() +
  # scale_x_continuous(trans='log2')+
  geom_smooth(method=lm) 

ggplot(scrn_usg_time, aes(small.scr.before.sleep, M)) + 
  geom_point() + 
  labs(title = "Sleep Quality vs Small Screen Usage") +
  theme_bw() +
  xlab("Off-screen Mimute Before Sleep") + 
  ylab("Sleep Quality(M)") + 
  # scale_x_continuous(trans='log2')+
  geom_smooth(method=lm) +
  xlim(0,60)

# c = c(1, 1)
# ggplot(scrn_usg_time, aes(c[1]*small.scr.before.sleep+c[2]*large.scr.before.sleep, M)) + 
#   geom_point() + 
#   labs(title = "Sleep Quality vs Screen Usage") +
#   theme_bw()  + 
#   # scale_x_continuous(trans='log2') + 
#   geom_smooth(method=lm) 

```


<!-- Heavy activity -->


<!-- ```{r, out.width="50%"} -->
<!-- ggplot(df_act, aes(heavy,M)) +  -->
<!--   geom_point() +  -->
<!--   labs(title = "Sleep Quality vs Heavy Activity") + -->
<!--   theme_bw()+  -->
<!--   geom_smooth(method=lm)  -->
<!-- ``` -->


```{r}
# df_pos = position %>% spread(position, total)
# df_pos['user'] = lapply(df_pos['user'],function(x) as.numeric(str_remove(x, "user_")))
# df_pos = df_pos %>% filter(user!=11) %>% arrange(user)
# df_pos['M'] = mydata$M
# df_pos$Inclinometer.Lying = as.numeric(df_pos$Inclinometer.Lying)
# df_pos$Inclinometer.Off = as.numeric(df_pos$Inclinometer.Off)
# df_pos$Inclinometer.Sitting = as.numeric(df_pos$Inclinometer.Sitting)
# df_pos$Inclinometer.Standing = as.numeric(df_pos$Inclinometer.Standing)
# df_pos.rcorr = rcorr(as.matrix(df_pos[-1]))
# df_pos.coeff = df_pos.rcorr$r
# df_pos.p = df_pos.rcorr$P
# 
# heatmap.2(x = df_pos.coeff, symm = TRUE, key = TRUE,labRow = abbreviate(rownames(df_pos.coeff), minlength = 15), labCol = abbreviate(colnames(df_pos.coeff), minlength = 15), col = palette, margins=c(8,10), srtCol=45, main = "heatmap", density.info="none", trace="none")
```


```{r, fig.show="hold", out.width="50%"}
# ggplot(df_pos, aes(Inclinometer.Off,M)) + geom_point() + labs(title = "Sleep Quality vs Time off") +theme_bw()
# ggplot(df_pos, aes(Inclinometer.Standing,M)) + geom_point() + labs(title = "Sleep Quality vs Time of Standing") +theme_bw()
# ggplot(df_pos, aes(Inclinometer.Sitting,M)) + geom_point() + labs(title = "Sleep Quality vs Time of Sitting") +theme_bw()
# ggplot(df_pos, aes(Inclinometer.Lying,M)) + geom_point() + labs(title = "Sleep Quality vs Time of Lying") +theme_bw()
```



### How do Biomarkers Affect Sleep Quality?
Research has shown that cortisol and melatonin are hormones produced by the body that regulate circadian rhythm and internal body clock. Melatonin eases you into sleep and helps you stay asleep. Its production is triggered by darkness and inhibitted by light. Cortisol helps you get up and keeps you awake. In summary, cortisol and melatonin have opposing rhythms. The prerequisites of a good sleep include having low cortisol and high melatonin levels before going to sleep and low melatonin and high cortisol levels after waking up.

We wanted to see if our metric is able to capture the presence of prerequisites or not. We plotted a bar graph of differences in concentration (before sleep - after sleep) for each of the hormones and users. Concentrations of melatonin run in $10^{-9}$ and so the differences are not small, given the scale. We saw that only three users have negative cortisol difference levels.

To simulateneously compare the differences with M, we generated a bar graph for sleep quality of all users. There isn't any "obvious" relationship between our metric and the hormones. However, our metric does seem to catch users like user 2 and 12 that fall on the two extremes. User 2 scored the lowest on sleep quality and has the highest positive differnece in cortisol levels in the cohort and the rhythm of melatonin is the same as cortisol's. User 12 is the only participant with opposing rhythms and also has the highest sleep quality among all.

There are users that don't follow a trend like user 1. One would think that user 1 would have low sleep quality given the differences but in fact ranks second in sleep quality. 

```{r}
# get the difference of chemicals per participant
before_sleep = saliva %>% 
  gather("Chemicals","Concentration", 3:4) %>% 
  filter(SAMPLES == "before sleep") %>% 
  arrange(user, Chemicals)

after_sleep = saliva %>% 
  gather("Chemicals","Concentration", 3:4) %>% 
  filter(SAMPLES == "wake up") %>% 
  arrange(user, Chemicals)

difference = data.frame(user = before_sleep$user, Chemicals = before_sleep$Chemicals, difference = after_sleep$Concentration - before_sleep$Concentration)

difference %>% ggplot(aes(x = factor(user,  levels = paste0("user_",c(1:22)), labels = c(1:22)), y = difference)) +
  geom_col(fill = 'lightblue', color = 'black') +
  scale_x_discrete(labels = abbreviate)+
  facet_grid(rows = vars(Chemicals), scales = "free") + 
  xlab("User ID") +
  ylab("Concentration") +
  ggtitle("Concentration Difference per User") +
  theme_bw()

mydata %>% ggplot(aes(x = user, y = M)) +
  geom_col(fill = 'lightblue', color = 'black') + 
  scale_x_continuous(breaks = c(1:22), labels = c(1:22)) +
  labs(x = "User ID", y = "Sleep Quality(M)", title = "Sleep Quality of each User") + 
  theme_bw()
```




```{r}

# difference['user.id'] = lapply(difference['user'], function(x) as.numeric(str_remove(x,"user_")))
# difference_temp <- difference[difference$user.id!=11,]
# difference_temp <- difference_temp[order(difference_temp$user.id),]
# sleep_quality_metric = mydata$M[mydata$user != 21]
# difference_temp$M <- unlist(lapply(sleep_quality_metric, function(x) c(rep(c(x),2))))
# 
# difference_temp %>% ggplot(aes(`difference`,`M`)) + geom_point() + facet_wrap(~Chemicals, scales = "free") + labs(x = "Difference in Concentration Before and After Sleep", y = "Sleep Quality (M)")

```

```{r}
# save workplace after data cleaning/for this stage.
save.image(file = "./data/RData/data_MMASH_results.RData")
```
