==> 06-interactive.Rmd <==
# Interactive component

```{r}
library(plotly)
library(ggplot2)
library(dplyr)
library(tidyr)
library(stringr)

load("./data/RData/data_MMASH_results.RData")

temp = difference %>% tidyr::spread(key = Chemicals, value = difference)
temp['user'] = lapply(temp['user'] , function(x)as.numeric(str_remove(x, "user_")))

df_plot = sleep_data %>% 
  select( M, In.Bed.Time, Out.Bed.Time) %>% 
  mutate(user = sapply(rownames(sleep_data),simplify =TRUE, function(x)as.numeric(stringr::str_remove(x, "user_")))) %>% 
  left_join(temp, by = 'user')

avg_saliva = saliva %>% group_by(user) %>% summarise(avg.Cortisol = mean(Cortisol.NORM), avg.Melatonin = mean(Melatonin.NORM))
avg_saliva['user'] = lapply(avg_saliva['user'] , function(x)as.numeric(str_remove(x, "user_")))

temp = df_plot['user']

df_plot = df_plot %>% 
  left_join(avg_saliva, by = 'user')

df_plot['user'] = temp
df_plot = df_plot %>% arrange(M)
readr::write_csv(df_plot,"./data/interactive_plot.csv")
```

Reader could explore the relation of biofactors to sleep quality by clicking the buttons. `Sleep Quality (M)` is the metric we defined. In this project, our main concern is how different factors affected sleep quality. Here, we focus on the difference of chemical levels before and after sleep. The key findings are present in the Results section. The `User ID` is arranged in ascending sleep quality order in convenience of comparing different plots. With the emphasis brought by the interactive transition, readers could easily tell the difference between patterns.

<meta charset="utf-8">
          
<!-- Load d3.js -->

<script src="https://d3js.org/d3.v7.js"></script>
<script src="https://colorbrewer2.org/export/colorbrewer.js"></script>

<!-- Add 2 buttons -->
<button onclick="update('M')">Sleep Quality (M)</button>
<button onclick="update('Cortisol.NORM')">Difference in Cortisol levels</button>
<button onclick="update('Melatonin.NORM')">Difference in Melatonin levels</button>

<!-- Create a div where the graph will take place -->
<div id="my_dataviz"></div>
<script>

// set the dimensions and margins of the graph
const margin = {top: 30, right: 30, bottom: 70, left: 60},
    width = 460 - margin.left - margin.right,
    height = 400 - margin.top - margin.bottom;
    
// append the svg object to the body of the page
var svg = d3.select("#my_dataviz")
  .append("svg")
    .attr("width", width + margin.left + margin.right)
    .attr("height", height + margin.top + margin.bottom)
  .append("g")
    .attr("transform", `translate(${margin.left},${margin.top})`);
    
// Initialize the X axis
const x = d3.scaleBand()
  .range([ 0, width ])
  .padding(0.2);
  
const xAxis = svg.append("g")
  .attr("transform", `translate(0,${height})`);
// Initialize the Y axis
const y = d3.scaleLinear()
  .range([ height, -1]);
  
const yAxis = svg.append("g")
  .attr("class", "myYaxis");
  
<!-- // Initialize the color scale -->
<!-- const colorScale = d3.scaleQuantize().range(colorbrewer.Blues[7]); -->

// text label for the x axis
svg.append("text")             
    .attr("transform",
          "translate(" + (width/2) + " ," + 
                         (height + margin.top + 20) + ")")
    .style("text-anchor", "middle")
    .text("User ID");
    
// text label for the y axis
svg.append("text")
    .attr("transform", "rotate(-90)")
    .attr("y", 0 - margin.left)
    .attr("x",0 - (height / 2))
    .attr("dy", "1em")
    .style("text-anchor", "middle")
    .text("Value");  

// A function that create / update the plot for a given variable:
function update(selectedVar) {

  // Parse the Data
  d3.csv("https://raw.githubusercontent.com/Shruti-Kaushal/MultilevelSleepAnalysis/main/data/interactive_plot.csv").then( function(data) {
  
  console.log(data);
  
  // X axis
  x.domain(data.map(d => d.user));
  xAxis.transition().duration(1000).call(d3.axisBottom(x));
  
  // Add Y axis
  y.domain([d3.min(data, d => +d[selectedVar]), d3.max(data, d => +d[selectedVar]) ]);
  yAxis.transition().duration(1000).call(d3.axisLeft(y));
  
  <!-- // Add color -->
  <!-- colorScale.domain([0, d3.max(data, d => d.M)]); -->
  
  // variable u: map data to existing bars
  const u = svg.selectAll("rect")
    .data(data);
  // update bars
  u.join("rect")
    .transition()
    .duration(1000)
      .attr("x", d => x(d.user))
      .attr("y", function(d) { return y(Math.max(0, d[selectedVar])); })
      .attr("width", x.bandwidth())
      .attr("height", function(d) { return Math.abs(y(d[selectedVar]) - y(0)); });
    
    
  });
}
// Initialize plot
update('M')
</script>

<font size="2">
Note: We tried to fix the problem of bar overflowing axis in the sleep quality plot. However, in order to show the negative scale in other plots, we still keep it as it is. We also noticed that sometimes the label will overlap with the label of y-axis. 
</font>
